+++ test/karma.conf.js
@@ -38,6 +38,8 @@ module.exports = function (config) {
             'bower_components/unchatbar-connection/app/scripts/controller/dialer.js',
             'bower_components/unchatbar-connection/app/scripts/controller/model-password.js',
             'bower_components/angular-scroll-glue/src/scrollglue.js',
+            'bower_components/angular-sanitize/angular-sanitize.js',
+            'bower_components/ng-emoticons/src/ng-emoticons.js',
             // endbower
             'bower_components/angular-mocks/angular-mocks.js',
             '!app/scripts/template.js',